refactor(BA-6554): drop AppConfigScopeType re-exports, import from common.data directly (review)
Per review: no re-export of AppConfigScopeType anywhere. Every consumer (manager
data/models/repositories/services/adapters, allow_list v2 DTO request/response, GQL types,
client CLI, tests) now imports it directly from ai.backend.common.data.app_config.types.
The DTO/data types modules keep the import only for their own filter/dataclass use and no
longer list it in __all__.
